### 2019-01-29 #1217 @whedon
[PRE REVIEW]: dfitspy
**Submitting author:** @astrom-tom (<a href="http://orcid.org/0000-0001-8385-3276">Romain Thomas</a>)
**Repository:** <a href="https://github.com/astrom-tom/dfitspy" target ="_blank">https://github.com/astrom-tom/dfitspy</a>
**Version:** v19.1.5
