### 2019-01-29 #1217 @danielskatz
@astrom-tom - I don't see any .bib file in your repo, and your reference is written manually, rather than being compiled.  Please refer to https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for an example.  Once you've made changes in your repository, enter ```@whedon generate pdf``` as a new comment here to attempt to rebuild the pdf from the md and bib files.
