### 2019-01-29 #1217 @danielskatz
You also need some real content in the readme, including the purpose of the software, what problems it can solve, a pointer to how it can be installed and tested, how others can contribute, etc.

See https://joss.readthedocs.io/en/latest/review_criteria.html, which states:
> A high-level overview of this documentation should be included in a README file (or equivalent).
