### 2019-01-29 #1217 @danielskatz
Also, I see a bunch of temporary files inside the repository that you might want to remove - those that end with `~`
