### 2019-01-29 #1217 @danielskatz
Once these changes are made, let me know and we'll move forward to assigning an editor
