+++ paper.bib
@@ -1,0 +1,15 @@
+@ARTICLE{Hanisch:2001,
+       author = {{Hanisch}, R.~J. and {Farris}, A. and {Greisen}, E.~W. and {Pence},
+        W.~D. and {Schlesinger}, B.~M. and {Teuben}, P.~J. and
+        {Thompson}, R.~W. and {Warnock}, A., III},
+        title = "{Definition of the Flexible Image Transport System (FITS)}",
+      journal = {\aap},
+     keywords = {INSTRUMENTATION: MISCELLANEOUS, METHODS: MISCELLANEOUS, TECHNIQUES: MISCELLANEOUS, ASTRONOMICAL DATABASES: MISCELLANEOUS},
+         year = 2001,
+        month = Sep,
+       volume = {376},
+        pages = {359-380},
+          doi = {10.1051/0004-6361:20010923},
+       adsurl = {https://ui.adsabs.harvard.edu/\#abs/2001A&A...376..359H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
