+++ paper.md
@@ -1,7 +1,7 @@
 ---
 title: 'dfitspy: a dfits|fitsort implementation in python'
 tags:
-  - fitsfiles
+  - fits files
   - Python
   - Astronomy
 authors:
@@ -12,6 +12,7 @@ affiliations:
  - name: European Southern Observatory, Av. Alonsode Córdov a3107, 7630355 Vitacura, Santiago, Chile
    index: 1
 date: 28 January 2019
+bibliography: paper.bib
 ---
 
 # Summary
@@ -28,7 +29,7 @@ software/eclipse/eug/eug/node8.html). The main limitation is that they are stand
 programs useable only in a terminal. They can not be used natively inside a another 
 program. 
 
-dfitspy was design to port the dfits and fitsort\footnote{https://www.eso.org/sci/software/eclipse/eug/eug/node8.html} capabilities to one of the most used 
+dfitspy was design to port the dfits and fitsort  capabilities to one of the most used 
 language used in science: python. It can be used as an executable (as the original 
 programs) but also, and it is its strenght, as a python module. Therefore fits file 
 can be selected in an efficient and fast way based on their metadata values inside 
