### 2019-01-31 e0ed1a2
paper corrections
