### 2019-02-01 #1217 @danielskatz
> Thank you for dealing with my submission. For the readme, there is only one link that goes to github pages (full documentation made with sphinx). Is that enough?

No, it's not.  This is what I wrote in https://github.com/openjournals/joss-reviews/issues/1217#issuecomment-458549216
