### 2019-02-02 #1217 @astrom-tom
A high-level overview of this documentation should be included in a README file (or equivalent).

What means 'or equivalent'?
In any case, I updated the readme and pointed also to the real online documentation.
