### 2019-02-03 #1217 @danielskatz
Equivalent is in case the software is not on github and in some other system where the initial file has some other name.
