### 2019-02-04 #1217 @danielskatz
@astrom-tom Regarding the readme, thanks for your updates, and, can you try to generalize it a bit? Think about how it will look to non-astronomers - will they have any idea what the software does? Or who might want to use it, and for what kind of research problems?
