### 2019-02-05 #1217 @astrom-tom
dear @danielskatz,

I updated the readme to briefly present the FITS file format and give a bit more details of what is dfitspy doing.
Let me know if this is ok.

Thank you!
