### 2019-02-05 8b1c234
wording changes
