+++ README.md
@@ -4,9 +4,9 @@ What is dfitspy?
 The full documentation can be found here: https://astrom-tom.github.io/dfitspy/build/html/index.html
 
 
-The FITS (Flexible Image Transport System) is one of the most used format in astronomy [https://fits.gsfc.nasa.gov/fits_documentation.html]. It is used to store a lot of different types of data such as 1D or 2D spectra, 3D data cubes. The FITS files are built with two components. The data themselves are stored as tables and contains any types of data. Then a header is built containing set of keywords-value pairs aiming at describing the data themselves. 
+The FITS (Flexible Image Transport System) is one of the most used formats in astronomy [https://fits.gsfc.nasa.gov/fits_documentation.html]. It is used to store a lot of different types of data, such as 1D or 2D spectra, 3D data cubes, etc. FITS files contain two components: a header contains a set of keywords-value pairs aiming at describing the data, and the data themselves are stored as tables and contain any type of data. 
 
-The present python module, dfitspy, is a small project that migrates the main dfits and fitsort capabilities to python [https://www.eso.org/sci/software/eclipse/eug/eug/node8.html]. It is a metadata searcher/displayer for FITS files. As dfits and fitsort, dfitspy is able to display in the terminal the result of a metadata search and is able to grep certain values of keywords inside large samples of files. Therefore it can be used directly with the command line interface. Nevertheless, dfitspy can be, and it is its strength, imported as a python module the user can use these functionnalities inside another python code or the python interpretor.  
+The present python module, dfitspy, is a small project that migrates the main dfits and fitsort capabilities to python [https://www.eso.org/sci/software/eclipse/eug/eug/node8.html]. It is a metadata searcher/displayer for FITS files. Similarly to dfits and fitsort, dfitspy is able to display the result of a metadata search in the terminal and is able to grep certain values of keywords in large samples of files. Therefore it can be used directly via the command line interface. Nevertheless, the strength of dfitspy is that it can be imported as a python module, and the user can use its functionality inside another python code or the python interpretor.  
 
 
 
