### 2019-02-05 #1217 @arfon
> 👋 @arfon - Can I pass this to you to edit? It also probably needs a check regarding the JOSS criteria about utility packages.

Yep, I may give this to our new editor. Regardless, I can take it from here.
