### 2019-02-07 #1217 @arfon
:wave: @astrom-tom - thanks for submitting this package to JOSS. Could you say a little more about how this submission meets the following [submission criteria](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements)?

> - The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

For this, it's not clear to me that this functionality isn't available elsewhere (e.g. in https://github.com/esheldon/fitsio or astropy).

Additionally:

> - The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications).

Have you considered contributing this to a larger community package such as `astropy.io.fits`? This feels like a more natural home for functionality like this than a small standalone package.
