### 2019-02-12 #1217 @astrom-tom
Dear @arfon,

Thank you for being the editor for my code.
To answer your questions:
- This package have been partially developed at the Very Large Telescope to ease the search and the display of fitsfiles metadata among large number of files. dfistpy primarily aims at being included in scripts where large number of fitsfiles have to be reviewed in an fast and efficient way (without looking at the data), which is the case at the telescope but also for programs that have a lot of data. Additionally, it can be used directly in the terminal as a stand-alone package for simple display in the terminal (a-la dfits+fitsort). It is worth noting that my supervisor allowed me to submit this package and the package will remain in active development.

-dfitspy is designed to do one thing only and has been optimized for it. Of course, with some work you can rewrite this for your own code (which is the case for, I guess, a lot of codes). I used fitsio and numpy modules as building blocks. I also tried using astropy but it seems that using fitsio was much faster when I tested with thousands of files (using the cfitsio wrapper directly included in the fitsio python module). I do not believe that including this in the astropy.io.fits is a good option.

Hope this answers your questions and do not prevent the review of this module.

Best,

R.
