### 2019-02-12 #1217 @arfon
> Hope this answers your questions and do not prevent the review of this module.

I think this is sufficient. Thanks for the additional background.

@taldcroft - would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
