### 2019-02-12 #1217 @taldcroft
@arfon - what's  the time scale for review?  Next 2-3 of weeks are tough, but after that I'd be happy to review this submission.
