### 2019-03-11 astrom-tom/dfitspy#3 @taldcroft
Following installation docs fails
The installation docs explicitly state that pip install will install the numpy and fitsio dependencies:
"""
The last dfitspy version is v19.1.5 and is available in the pypi test repository. To install it:
```
pip install -i https://test.pypi.org/simple/ dfitspy --user
```
Using this command will allow you not to have to install any other package. Pip will install what is missing for you.
"""
Unfortunately `setup.py` imports the package itself (`import dfitspy.__info__ as dfits`), so in a clean environment with no numpy / fitsio already installed one gets an exception:
```
(dfitspy) monkeys$ pip install -i https://test.pypi.org/simple/ dfitspy
Looking in indexes: https://test.pypi.org/simple/
Collecting dfitspy
  Downloading https://test-files.pythonhosted.org/packages/00/e3/70ef5aa075dbe22aab5d99ecb855ec3085ae742ae5290b671085c9ee2a6e/dfitspy-19.1.3.tar.gz (7.5MB)
    100% |████████████████████████████████| 7.5MB 6.7MB/s
    Complete output from command python setup.py egg_info:
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/k1/2ktyqf0x2n7_fzvdnvljq2240000gq/T/pip-install-p4qswz15/dfitspy/setup.py", line 2, in <module>
        import dfitspy.__info__ as dfits
      File "/private/var/folders/k1/2ktyqf0x2n7_fzvdnvljq2240000gq/T/pip-install-p4qswz15/dfitspy/dfitspy/__init__.py", line 6, in <module>
        from dfitspy.readfits import read_fitsfile, get_all_keyword, keywords_in_file, dfitsort
      File "/private/var/folders/k1/2ktyqf0x2n7_fzvdnvljq2240000gq/T/pip-install-p4qswz15/dfitspy/dfitspy/readfits.py", line 35, in <module>
        import numpy
    ModuleNotFoundError: No module named 'numpy'
```

Some options are
- Update the installation docs to provide the commands to install numpy and fitsio
- Fix `setup.py` so it can be run without the runtime dependencies already installed
  - Make the package importable (though not runnable) without dependencies
  - Use something like setuptools_scm to single-source the version from git tag without needing to import the package in `setup.py`.  (See also https://github.com/pypa/setuptools_scm/issues/273 for some useful discussion).
