### 2019-03-11 astrom-tom/dfitspy#4 @taldcroft
Minor install docs issue
The install docs say `pip install dfitspy.X.Y.Z.tar.gz --user` but the filename has a `-` not `.` before the version `dfitspy-X.Y.Z.tar.gz`.
