### 2019-03-11 astrom-tom/dfitspy#5 @taldcroft
`dfitspy --docs` always opens local docs
Even with a functioning internet connection, asking for docs always gives me the local version:
```
(dfitspy) monkeys$ ping google.com
PING google.com (172.217.12.206): 56 data bytes
64 bytes from 172.217.12.206: icmp_seq=0 ttl=55 time=8.902 ms
64 bytes from 172.217.12.206: icmp_seq=1 ttl=55 time=8.572 ms
64 bytes from 172.217.12.206: icmp_seq=2 ttl=55 time=8.612 ms
^C
--- google.com ping statistics ---
3 packets transmitted, 3 packets received, 0.0% packet loss

(dfitspy) monkeys$ dfitspy --docs
No internet connection detected, open local documentation
/usr/bin/open
```
Not a huge problem, but the docs suggest otherwise.
