### 2019-03-11 astrom-tom/dfitspy#6 @taldcroft
--grep is not really grepping and raises exception for no matching
I would expect that the `--grep` option accepts regular expressions or at least matches a string that contains the grep value.  Right now `--grep` only succeeds for an exact match of the entire string, and requires at least one match or there is a crash.

```
[DFITSPY INFO]> Current directory: /Users/aldcroft/git/dfitspy/dfitspy/tests
[DFITSPY INFO]> 5 fits files will be considered

filename  	YEAR  	PLACE      	AUTHOR
----------	------	-----------	-------------
test.fits 	2018  	ESO Paranal	Romain Thomas
test2.fits	2018  	ESO Paranal	Romain Thomas
test3.fits	2018  	ESO Paranal	Romain Thomas
test4.fits	2018  	ESO Paranal	Romain Thomas
test5.fits	201810	ESO Paranal	Romain Thomas

[DFITSPY INFO]> 5 files used in output
```
```
(dfitspy) monkeys$ dfitspy -f test*.fits -k YEAR,PLACE,AUTHOR --grep 2018

[DFITSPY INFO]> Current directory: /Users/aldcroft/git/dfitspy/dfitspy/tests
[DFITSPY INFO]> 5 fits files will be considered

filename  	YEAR	PLACE      	AUTHOR
----------	----	-----------	-------------
test.fits 	2018	ESO Paranal	Romain Thomas
test2.fits	2018	ESO Paranal	Romain Thomas
test3.fits	2018	ESO Paranal	Romain Thomas
test4.fits	2018	ESO Paranal	Romain Thomas

[DFITSPY INFO]> 4 files used in output
```
```
(dfitspy) monkeys$ dfitspy -f test*.fits -k YEAR,PLACE,AUTHOR --grep 2017

[DFITSPY INFO]> Current directory: /Users/aldcroft/git/dfitspy/dfitspy/tests
[DFITSPY INFO]> 5 fits files will be considered

Traceback (most recent call last):
  File "/Users/aldcroft/anaconda3/envs/dfitspy/bin/dfitspy", line 10, in <module>
    sys.exit(main())
  File "/Users/aldcroft/anaconda3/envs/dfitspy/lib/python3.6/site-packages/dfitspy/__main__.py", line 132, in main
    dp.dfitsort_view(allvalues)
  File "/Users/aldcroft/anaconda3/envs/dfitspy/lib/python3.6/site-packages/dfitspy/display.py", line 58, in dfitsort_view
    maxsize_filename = max([len(os.path.basename(i)) for i in filenames])
ValueError: max() arg is an empty sequence
```
```
(dfitspy) monkeys$ dfitspy -f test*.fits -k YEAR,PLACE,AUTHOR --grep Thomas

[DFITSPY INFO]> Current directory: /Users/aldcroft/git/dfitspy/dfitspy/tests
[DFITSPY INFO]> 5 fits files will be considered

Traceback (most recent call last):
  File "/Users/aldcroft/anaconda3/envs/dfitspy/bin/dfitspy", line 10, in <module>
    sys.exit(main())
  File "/Users/aldcroft/anaconda3/envs/dfitspy/lib/python3.6/site-packages/dfitspy/__main__.py", line 132, in main
    dp.dfitsort_view(allvalues)
  File "/Users/aldcroft/anaconda3/envs/dfitspy/lib/python3.6/site-packages/dfitspy/display.py", line 58, in dfitsort_view
    maxsize_filename = max([len(os.path.basename(i)) for i in filenames])
ValueError: max() arg is an empty sequence
```
