### 2019-03-11 astrom-tom/dfitspy#7 @taldcroft
Provide links to dfits and fitsort docs or description
Since this tool is intended to provide a Python equivalent to `dfits` and `fitsort`, it would be helpful to provide links in the `dfitspy` documentation to something describing the originals.
