### 2019-03-11 astrom-tom/dfitspy#8 @taldcroft
Provide some performance measure and compare with `astropy`
The documentation says that dfitspy "allows you to find common values in large amount of fits files in a very fast way".

Can you provide some context for what "very fast" means?  Can you compare a grepping operation on a large list of files to doing the same with `astropy.io.fits`?  I would expect `astropy.io.fits` to be quite a bit slower given the design choices there, so it might highlight the utility of this package.
