### 2019-03-11 astrom-tom/dfitspy#9 @taldcroft
Provide API docs
Although the function help functions are mostly including in the narrative docs, it would be good to provide standard sphinx API docs.
