### 2019-03-11 astrom-tom/dfitspy#10 @taldcroft
Provide a link to the github repo in docs
The main docs page does not include a link back to the github repo, making it more difficult for people to contribute or report issues.
