### 2019-03-11 #1249 @taldcroft
Minor comments on the article proof:
- `fits` is an acronym and should always be capitalized as `FITS`.
- `usefull` => `useful`.
- `used natively inside a another program` (extra `a`)
- `strenghth` => `strength`
- Not sure of JOSS standards, but I believe `Python` should always be capitalized.
- `Therefore fits file can` => `Therefore FITS files can` (plural `files`)
- There is a space in the URL for the `www.eso.org/...` link.  Note also that this link goes to a page that provides man page links for the original `dfits` and `fitsort` standalone tools, but both those links are broken.  Can you dig up any working links for the `difts` and `fitsort` man pages?
