### 2019-03-11 astrom-tom/dfitspy#11 @taldcroft
Typos in docs
There are a number of typos in the docs that could be fixed with an English spell checker.  It would be worth doing this to make the package seem a bit more polished.
