### 2019-03-11 #1249 @taldcroft
@arfon @astrom-tom - Review completed.

I opened a number of issues that I think should be addressed (by fixing or discussion).  These issues  are the source of unchecked boxes in the review checklist.

- https://github.com/astrom-tom/dfitspy/issues/3
- https://github.com/astrom-tom/dfitspy/issues/4
- https://github.com/astrom-tom/dfitspy/issues/5
- https://github.com/astrom-tom/dfitspy/issues/6
- https://github.com/astrom-tom/dfitspy/issues/7
- https://github.com/astrom-tom/dfitspy/issues/8
- https://github.com/astrom-tom/dfitspy/issues/9
- https://github.com/astrom-tom/dfitspy/issues/10
- https://github.com/astrom-tom/dfitspy/issues/11

Overall I see the utility in this package as a standalone package, separate from `astropy` or `fitsio`.  (For astropy there is a GPL / BSD license problem anyway).  When we've ironed the issues I opened then I'll be happy to recommend inclusion to JOSS.  :smile:
