### 2019-03-11 astrom-tom/dfitspy#5 @astrom-tom
Fixed, this was due to a missing package ('socket').
