### 2019-03-11 af494f8
issue 4,5 and 6
