### 2019-03-12 astrom-tom/dfitspy#4 @astrom-tom
fixed, added as well the description of X, Y and Z so people can spot the last version more easily.
