### 2019-03-12 astrom-tom/dfitspy#10 @astrom-tom
Fixed, Two links have been added in the from page of the documentation (one at the beginning of the first paragraph, and one at the end of the contribute section)
