### 2019-03-12 astrom-tom/dfitspy#7 @astrom-tom
Dear @taldcroft ,

I could find two links to the manpage:

ftp://134.171.42.54/scisoft/scisoft4/linux/redhat9/eclipse/eclipse/man/html1/dfits.html
ftp://134.171.42.54/scisoft/scisoft4/linux/redhat9/eclipse/eclipse/man/html1/fitsort.html

But I am really not sure about the sustainability of these links. I could not find more stable so I was thinking I could copy/paste the manpages and add them to the documentation. What do you think?
