### 2019-03-12 astrom-tom/dfitspy#11 @astrom-tom
Dear @taldcroft,

I used an English spell checker to correct typos. I found quite a lot and corrected them.
