+++ dfitspy/docs/docs/logo.txt
@@ -1,0 +1,10 @@
+      _  __ _ _       ____        
+   __| |/ _(_) |_ ___|  _ \ _   _ 
+  / _` | |_| | __/ __| |_) | | | |
+ | (_| |  _| | |_\__ \  __/| |_| |
+  \__,_|_| |_|\__|___/_|    \__, |
+     R. Thomas 2018-19      |___/ 
+
+
+
+
