+++ apidoc/modules.rst
@@ -1,0 +1,7 @@
+dfitspy
+=======
+
+.. toctree::
+   :maxdepth: 4
+
+   dfitspy
