### 2019-03-12 astrom-tom/dfitspy#9 @astrom-tom
Dear @taldcroft ,

I added a page in the doc with the sphinx API docs of all the useful functions.
