### 2019-03-12 astrom-tom/dfitspy#3 @astrom-tom
The problem was to get the metadata from the __info__.py in the package module. I wrote these metadata in the setup.py so it is not necessary to import the package in setup.py. This should fix the problem.
