+++ apidoc/modules.rst
@@ -1,7 +1,0 @@
-dfitspy
-=======
-
-.. toctree::
-   :maxdepth: 4
-
-   dfitspy
