+++ apidoc/dfitspy.rst
@@ -1,54 +1,0 @@
-dfitspy package
-===============
-
-Submodules
-----------
-
-dfitspy.cli module
-------------------
-
-.. automodule:: dfitspy.cli
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-dfitspy.display module
-----------------------
-
-.. automodule:: dfitspy.display
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-dfitspy.get\_files\_and\_keys module
-------------------------------------
-
-.. automodule:: dfitspy.get_files_and_keys
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-dfitspy.readfits module
------------------------
-
-.. automodule:: dfitspy.readfits
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-dfitspy.tests module
---------------------
-
-.. automodule:: dfitspy.tests
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: dfitspy
-    :members:
-    :undoc-members:
-    :show-inheritance:

[…]
