### 2019-03-12 astrom-tom/dfitspy#6 @astrom-tom
I fixed the crash (no match with the grep --> no list displayed).
I also improved the grepping. Now if one keyword value (among the requested keyword), contain the grepping value (it can be that the grepping value is exaclty the keyword value or the grepping value can be contained in the keyword value)
