### 2019-03-12 astrom-tom/dfitspy#8 @astrom-tom
This I will need some time because it means I have to rewrite the FITS reading  function in astropy :D.
