### 2019-03-12 astrom-tom/dfitspy#8 @taldcroft
@astrom-tom - the other way out is to just rephrase the docs as "allows you to find common values in a large number of FITS files in an efficient and convenient way".  That is definitely true: efficient because it is only reading the headers, and convenient because it manages all the low-level calls and provides a nice higher-level API.
