### 2019-03-13 astrom-tom/dfitspy#11 @taldcroft
Looks better!
