### 2019-03-13 astrom-tom/dfitspy#9 @taldcroft
Looks good.
