### 2019-03-13 astrom-tom/dfitspy#7 @taldcroft
Yes, since dfitspy has a clear heritage, it would be worth it to just paste those man pages as plain (pre-formatted) text in to a separate RST doc (e.g. `dfits and fitsort man pages`) that is linked in the top Table of Contents.
