### 2019-03-13 astrom-tom/dfitspy#3 @taldcroft
The latest version 19.3.2 (from the tarball) does install in a clean environment.  However, the install page and the PyPI link is still pointing at 19.1.5.
