### 2019-03-13 astrom-tom/dfitspy#5 @taldcroft
Looks good now.
