### 2019-03-13 astrom-tom/dfitspy#6 @taldcroft
Much better!
