+++ dfitspy/docs/source/conf.py
@@ -26,7 +26,7 @@ author = 'R. Thomas'
 # The short X.Y version
 version = '19.3'
 # The full version, including alpha/beta/rc tags
-release = '19.3.2'
+release = '19.3.3'
 
 
 # -- General configuration ---------------------------------------------------
