### 2019-03-13 astrom-tom/dfitspy#8 @astrom-tom
I rephrased as proposed.
