### 2019-03-13 astrom-tom/dfitspy#7 @astrom-tom
I added a separate page in the documentation with the fitsort and dfits manpage (also provided the links)
