### 2019-03-13 astrom-tom/dfitspy#3 @astrom-tom
I release the last version, 19.3.3, in the main pypi repository. A simple `pip install dfitspy` is now enough to install that version
