### 2019-03-13 #1249 @astrom-tom
@taldcroft, @arfon  The remaining issues have been adresses and the code have been updated and move to the main pypi repository. The new version is 19.3.3.
