+++ README.md
@@ -19,9 +19,9 @@ is not made for that (see fitsio and astropy). It works only on the header and a
 Installation?
 =============
 
-The last dfitspy version is v19.1.5 and is available in the pypi test repository. To install it::
+The last dfitspy version is v19.3.3 and is available in the pypi test repository. To install it::
 
-     pip install -i https://test.pypi.org/simple/ dfitspy --user
+     pip install dfitspy --user
 
 Using this command will allow you not to have to install any other package. Pip will install what is mis
 sing for you.
