### 2019-03-13 #1249 @taldcroft
All good now, I've reviewed and closed the outstanding issues and have now checked off the remaining review boxes.

From my perspective I'm pleased to recommend to @arfon and the editors that `dfitspy` be accepted to JOSS.

:tada:
