### 2019-03-13 #1249 @arfon
Thanks @taldcroft!

@astrom-tom - your paper is currently very short. Do you think it would be possible to add a little more detail for example discussing the need for this software and some examples of usage? https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
