+++ paper.md
@@ -46,9 +46,8 @@ Examples:
 \item The terminal interface (a-la-dfits|fitsort):\\
 ``dfitspy -f Test_data/* -k OBJECT,LST,ESO.OBS.ID --grep 79056.26 --grep STD,TELLURIC``\\
 This command will search is all the FITS file contained in the \textit{Test_data} directory. Dfitspy will search for three keywords in the header: OBJECT, LST and ESO OBS ID. Finally, the terminal will display only the file where both 79056.26 and STD,TELLURIC are in the keyword values.
-
 \item The python module:\\
-````
+``fwdef``
 \end{itemize}
 
 
