+++ paper.md
@@ -62,7 +62,7 @@ This command will search is all the FITS file contained in the \textit{Test_data
 
 
 # dfitspy as a Python module
-To be used as a Python module, dfitspy must be imported. Then a set of command have to be used in order to produce the final list of filenames/keywords/values. In short, two main commands must be used:
+To be used as a Python module, dfitspy must be imported. Then a set of command have to be used in order to produce the final list of filenames/keywords/values. In short, three main commands must be used:
 
 Import the module:\
 ``import dfitspy``
