+++ paper.md
@@ -24,28 +24,29 @@ any types of data. A header is built containing set of keywords-value pairs aimi
 the data themselves.
 
 Accessing and displaying metadata inside FITS files is important in order to get an overview
-of what is inside these files. It is particularly useful when dealing with large amount
-of files at once. Tools have been already publicly available for years with the dfits 
-and fitsort algorithms (the documentation is available here 
+of their content properties without having to read the data themselves. 
+It is particularly useful when dealing with large amount of files at once. 
+Tools have been already publicly available for years with the dfits and fitsort algorithms 
+(the documentation is available here 
 https://www.eso.org/sci/software/eclipse/eug/eug/node8.html). The main limitation is 
 that they are stand-alone programs useable only in a terminal. They can not be used natively 
 inside another program. 
 
-The python module presented in this paper, dfitspy, is a project that migrates the main dfits 
+The python module presented in this paper, ``dfitspy``, is a project that migrates the main dfits 
 and fitsort capabilities to python. It is a metadata searcher/displayer for FITS files. 
-As dfits and fitsort, dfitspy is able to display in the terminal the result of a metadata 
+As dfits and fitsort, ``dfitspy`` is able to display in the terminal the result of a metadata 
 search and is able to grep certain values of keywords inside large samples of files. 
 Therefore it can be used directly with the command line interface. Nevertheless, 
-dfitspy can be, and it is its strength, imported as a python module and the user can 
+``dfitspy`` can be, and it is its strength, imported as a python module and the user can 
 use these functionnalities inside another python code or the python interpretor.
 
 
 # dfitspy as a terminal command
-A command line interface has been included in dfitspy so it can be used as a Terminal command. A typical command is:\
+A command line interface has been included in ``dfitspy`` so it can be used as a Terminal command. A typical command is:\
 
 ``dfitspy -f Test_data/* -k author,number,type --grep 2dspec``\
 
-This command will search is all the FITS file contained in the \textit{Test_data} directory. Dfitspy will search for three keywords in the header: author, time and type. Finally, the terminal will display only the file where both 79098.26 and STD,TELLURIC are in the keyword values. The terminal output is similar to the dfits|fitsort combination. It displays in a column fashion, each file with the requested keyword and value for each keyword.
+This command will search in all the FITS file present in the \textit{Test_data} directory. ``dfitspy`` will search for three keywords in the header: author, time and type. Finally, the terminal will display only the file where ``2dspec`` is in the requested keyword values. The terminal output is similar to the dfits|fitsort combination. It displays, in a column fashion, each file with the requested keyword its corresponding values:
 
 \newpage
 
@@ -60,7 +61,7 @@ This command will search is all the FITS file contained in the \textit{Test_data
 
 
 # dfitspy as a Python module
-To be used as a Python module, dfitspy must be imported. Then a set of command have to be used in order to produce the final list of filenames/keywords/values. In short, three main commands must be used:
+To be used as a Python module, ``dfitspy`` must be imported. Then a set of command have to be used in order to produce the final list of filenames/keywords/values. In short, three main commands must be used:
 
 Import the module:\
 ``import dfitspy``
@@ -79,7 +80,7 @@ The final output is stored as a dictionnary of files for which each keywords/val
 
 # Availability
 
-dfitspy is a GPL licensed software and the source code is available at https://github.com/astrom-tom/dfitspy. The full documentation is available at https://astrom-tom.github.io/dfitspy/build/html/index.html .
+``dfitspy`` is a GPL licensed software and the source code is available at https://github.com/astrom-tom/dfitspy. The full documentation is available at https://astrom-tom.github.io/dfitspy/build/html/index.html .
 
 # Acknowledgements
 
