### 2019-03-13 #1249 @astrom-tom
Dear @arfon

First of all, sorry for the mess with all the pdf compilations. I reworked on the paper and included more information and gave simple examples.
