### 2019-03-13 #1249 @arfon
@astrom-tom - I've made a small correction here: https://github.com/astrom-tom/dfitspy/pull/12
