### 2019-03-13 #1249 @astrom-tom
Here it is: DOI: 10.5281/zenodo.2592698
One question, can I put dfitspy in ASCL?
