### 2019-03-14 ada0207
paper final
