### 2019-03-14 #1249 @arfon
> @arfon Here it is: DOI: 10.5281/zenodo.2592698
👍

@astrom-tom - before I proceed with accepting this paper can you confirm you've finished iterating on the paper?

> One question, can I put dfitspy in ASCL?

Definitely, although you probably want to make sure you're filling in the 'preferred citation' field to make sure that authors are citing the thing you want to be cited (presumably this paper). See for example Astropy's entry: http://ascl.net/1304.002
