### 2019-03-14 #1249 @astrom-tom
@arfon Yes, the paper is ok for me and I will not modify it.
As for ASCL I will wait that this paper is ready on ADS.
