### 2019-03-14 #1249 @astrom-tom
Thank you so much for your work @arfon and @taldcroft. It is really awesome!
(I just signed up for being a reviewer :D ).

@arfon One very last question, when will it be available in ADS? Is there any timeline?
