### 2019-03-14 #1249 @arfon
> (I just signed up for being a reviewer :D ).

Thanks!

> @arfon One very last question, when will it be available in ADS? Is there any timeline?

I think ADS typically index us every few weeks. It's not immediate.
