### 2019-03-14 #1249 @astrom-tom
@arfon Thanks a lot!
For the future version, should I resubmit them or can I just cite the paper for the version that was just accepted?
