### 2019-03-14 #1249 @arfon
> For the future version, should I resubmit them or can I just cite the paper for the version that was just accepted?

Future versions of this package? If there have been major updates then it's possible to resubmit and publish another JOSS paper but this has only happened for a couple of packages in the past and the changes were very large. Basically unless you feel like you need to publish a new paper, I'd encourage people to cite this paper.
